Fix incorrect army data loaded on the Unit page
The Unit page wasn't taking into account the army composition of the list when
loading army data (the units).

Fix #178
