+++ src/pages/unit/Unit.js
@@ -24,9 +24,10 @@ import { nameMap } from "../magic";
 import { editUnit, removeUnit, duplicateUnit } from "../../state/lists";
 import { setArmy } from "../../state/army";
 import { openRulesIndex } from "../../state/rules-index";
+import { getArmyData } from "../../utils/army";
 import { useLanguage } from "../../utils/useLanguage";
 import { updateLocalList } from "../../utils/list";
-import { updateIds, getRandomId } from "../../utils/id";
+import { getRandomId } from "../../utils/id";
 import { normalizeRuleName } from "../../utils/string";
 import { getUnitName, getUnitOptionNotes } from "../../utils/unit";
 
@@ -436,16 +437,12 @@ export const Unit = ({ isMobile, previewData = {} }) => {
         url: `games/${list.game}/${list.army}`,
         onSuccess: (data) => {
           dispatch(
-            setArmy({
-              lords: updateIds(data.lords),
-              heroes: updateIds(data.heroes),
-              characters: updateIds(data.characters),
-              core: updateIds(data.core),
-              special: updateIds(data.special),
-              rare: updateIds(data.rare),
-              mercenaries: updateIds(data.mercenaries),
-              allies: updateIds(data.allies),
-            })
+            setArmy(
+              getArmyData({
+                data,
+                armyComposition: list.armyComposition || list.army,
+              })
+            )
           );
         },
       });
